Write the accounts DB on a timer instead of player triggered events
Writing the DB is resource intensive and should be done as little as
possible. Putting it on a timer is intended to prevent buggy/malicious
clients from consuming large amounts of resources. This patch may
cause minor inconsistancies between the login and character server
databases, as they are both doing delayed writes, probably on different
schedules. I'm unable to see any huge consequences to this.