tmwa-admin should no longer log passwords in cleartext
This is ugly. I'd rather X-out the password in the buffer while in the
addaccount function and THEN log the buffer afterwards, but I can't
figure out how to do that with all this weird string typing. I guess
it's a "good thing" I can't even figure out how to mash memory outside
of the local function?