Fix NPC script softlock
When an NPC called destroy; from an attached script, npc_id, and
associated state, was not properly cleared. This meant that the
server considered the PC to still be in conversation, preventing
many actions including walking, but with an invalid NPC.

This clears state for the specific case of the a PC attached to
the script instance that called destroy. It is still not safe to
call destroy; if other players may be attached to it.