Improve login-server logging
Sends many important messages to both stdout and to log, rather than one
or the other, which required often time consuming reconstruction.

Change the format of many logs so that when logging the IP of the
char-server, rather than the IP of the source of the request, it should
be clearer that this is what is meant.

Log the IP in many more cases, including for an incorrect password in a
failed login attempt, in a format suitable for fail2ban or other
anti-bruteforce measures.

Fix many miscellaneous grammatical or spelling mistakes.