Mages Quality Of Life, part 2: This commit lands fix for infamous AFK-kill summons exploit.
Its what essentially caused summoner spells in cities to be disabled.

This is not free lunch: it comes at expense of little fun where
mob would bite bad master hitting it. But its been minor fun or
cosmetics while spells restriction in cities and so on is annoying
setback for mages.

So if choosing between evil and even more evil, guess...

Technically, server's bug is: it too eager to reattach some things to
wrong RID or something like this. Its what causes aggroed mob to
lock on really wrong random targets and AFK-kill them.