strs: Storage class consistency
fwd provides forward declarations of using `class`, but these are defined
in strs.hpp using struct. Only a problem if you both use an explicit `struct`
and use link time optimization - the warning about violating ODR is emitted
during link time with -flto, there is no warning in an isolated compilation
unit.

I do not fully understand why yet.
See the following "never-expire" pastebin of a diff that attempts to create
an upper bound for a a more isolated test case, and fails to trigger the ODR
warning:
https://0bin.net/paste/jDz63gch#DM2x09aMX+iaIAaFWI-WmeSyyKto5F+vZzSniHoHfI6