Fix error messages in `countitem` and `end` built-ins
Function: builtin_countitem

The error msg using nameid made no sense as it is output only when there is no nameid info.
Changed to "no item ID".

Function: builtin_end

The error msg "close in a callfunc or callsub", refers to the wrong function.
Changed to msg "end in a callfunc or callsub".