qq003 camera, rid and player options
patch qq003_camera
[quietlyquietly]

Function: builtin_camera
The "rid" and "player" options were broken due to later conflicting if stmts.
Moved the HARG(1) and HARG(2) tests earlier so that immediate handling
of options is possible.
Remove "rel" boolean, as it was used only for "relative", and all it was doing could be done
by moving that special clif_npc_action to the "relative" clause and
doing an immediate return.
The if stmts that parse the options are made to be exclusive, each clause handles one option.
Each of them sets the var "bl" differently, and without conflict.

Co-authored-by: Thorbjørn Lindeijer <bjorn@lindeijer.nl>
Approved-by: Hello TMW <hello@themanaworld.org>