Readme: Rework distributor instructions
This can be left to the end of the Readme. It's a lot of
information that can be off-putting, and on the happy path
it shouldn't be required.

Bump version numbers to what we've actually been using
or develop with.

Remove snark about versions with problems. Rephrase a few other
lines in friendlier ways.

Not handled yet: Wiki's How to Develop references rEvolt/Evolved,
rather than tmwa. tmwa should either link to an older revision
or to an article for tmwa specific setup.