Call mob_stop_walking after applying damage
The `mob_stop_walking` triggers message 0x0078 to client, which includes
the mob health information. By this time the client has already received
the attack message 0x008a sent from `clif_damage` which was called
earlier in `battle_weapon_attack`, so it's a little odd to receive
outdated hp just after the attack message.

Fixed this by applying the damage earlier. As far as I could see the hp
value wasn't used anywhere in between so this should be safe.